Split out qRegisterInfo tests
